+++ utils/src/copilotEntry.ts
@@ -3,10 +3,5 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IActionContext } from "../..";
-
-const copilotUserInputCanaryKey = '_copilotUserInput';
-
-export function isCopilotUserInput(context: IActionContext): boolean {
-    return !!(context as unknown as Record<string, unknown>)[copilotUserInputCanaryKey];
-}
+export * from './copilot/copilot';
+export { CopilotUserInput, markAsCopilotUserInput } from './userInput/CopilotUserInput';
