+++ utils/src/index.ts
@@ -9,7 +9,6 @@ export { createAzExtLogOutputChannel, createAzExtOutputChannel } from './AzExtOu
 export * from './AzExtTreeFileSystem';
 export * from './callWithTelemetryAndErrorHandling';
 export { activityErrorContext, activityFailContext, activityFailIcon, activityInfoContext, activityInfoIcon, activityProgressContext, activityProgressIcon, activitySuccessContext, activitySuccessIcon } from './constants';
-export * from './copilot/copilot';
 export * from './copilot/installCopilotCli';
 export * from './createApiProvider';
 export { createExperimentationService } from './createExperimentationService';
@@ -53,7 +52,8 @@ export * from './tree/v2/createGenericElement';
 export * from './tree/v2/TreeElementStateManager';
 export * from './userInput/AzExtUserInput';
 export * from './userInput/AzExtUserInputWithInputQueue';
-export * from './userInput/CopilotUserInput';
+
+
 export * from './utils/activityUtils';
 export * from './utils/AzExtFsExtra';
 export * from './utils/AzureResourceIdTelemetry';
